Remove classes from Utils and move to separate files
