+++ src/main/resources/messages/RadicleBundle.properties
@@ -130,3 +130,4 @@ issueCreated=Created : {0}
 stateChangeError=There was an error updating state
 assignersChangeError=There was an error updating assigners
 tagChangeError=There was an error updating tag
+issueOverview=Issue Overview
